dealing with error when trying to delete a category that has videos associated with it

+++ src/main/java/com/gabrielpf/alurabackendchallange/service/CategoryService.java
@@ -11,16 +11,21 @@ import com.gabrielpf.alurabackendchallange.controller.form.CategoryCreateForm;
 import com.gabrielpf.alurabackendchallange.controller.form.VideoCreateForm;
 import com.gabrielpf.alurabackendchallange.dto.CategoryDto;
 import com.gabrielpf.alurabackendchallange.exception.DataAlreadyExistsException;
+import com.gabrielpf.alurabackendchallange.exception.EntityCannotBeDeletedException;
 import com.gabrielpf.alurabackendchallange.model.Category;
+import com.gabrielpf.alurabackendchallange.model.VideoCategory;
 import com.gabrielpf.alurabackendchallange.repository.CategoryRepository;
+import com.gabrielpf.alurabackendchallange.repository.VideoCategoryRepository;
 
 @Service
 public class CategoryService {
 
     private final CategoryRepository repository;
+    private final VideoCategoryRepository videoCategoryRepository;
 
-    public CategoryService(CategoryRepository repository) {
+    public CategoryService(CategoryRepository repository, VideoCategoryRepository videoCategoryRepository) {
         this.repository = repository;
+        this.videoCategoryRepository = videoCategoryRepository;
     }
 
     public List<CategoryDto> findAll() {
@@ -43,6 +48,10 @@ public class CategoryService {
     }
 
     public void delete(UUID id) {
+        var videoCategory = videoCategoryRepository.findByCategoryId(id);
+        if (videoCategory.isPresent())
+            throw new EntityCannotBeDeletedException(VideoCategory.class, "id", "There are videos still attached to this category");
+
         repository.deleteById(id);
     }
 }
